- moved externs to top of some files - tidied up whitespace - removed some commented out code - removed "/* for win95 net game support */" comments - marine HUD motion tracker renders incorrectly. is too small. testing HUD resize, will fix - changed some BOOL types to bool - test code for stars rendering in place, stars will not appear. will fix - moved effectID_t and shaderID_t typedefs into render.h - added Tex_AddExistingTexture() - commented out entire awPnmld.cpp file as it's not used (will remove from SVN) - specular is enabled again but might not work correctly in all cases yet - changed _D3DTVERTEX struct members to change sx, sy, sz to x, y, z and tu and tv to u and v - as above, changed incorrect name "_D3DTVERTEX" to _D3DLVERTEX (changed 'T' to 'L' for lit :) ) - changed some shader ID types in d3_func.h from uint32_t to correct effectID_t - added aspect ratio float to d3d struct - allow DrawFmvFrame() to handle non power of 2 padded textures - removed unused DrawHUDQuad() - updated xbox files to match above changes (hopefully) - in awTexLd.cpp, not using malloc for d3d_texture or buffer in CreateTexture() - using c++ 'new' now
Imported-from: https://subversion.assembla.com/svn/AvP/branches/workbranch/trunk@421

+++ 3dc/avp/scream.cpp
@@ -71,8 +71,8 @@ void CharacterSoundEffects::LoadSounds(const char* filename, const char* directo
 	}
 
 	char* buffer;
-	int file_size;
-	unsigned long bytes_read;
+	uint32_t file_size;
+	DWORD bytes_read;
 
 	file_size = GetFileSize(file, 0);
 	buffer = new char[file_size+1];
@@ -94,13 +94,13 @@ void CharacterSoundEffects::LoadSounds(const char* filename, const char* directo
 	voice_types = (ScreamVoiceType*) PoolAllocateMem(num_voice_types * sizeof(ScreamVoiceType));
 	
 	char wavpath[200];
-	strcpy(wavpath,directory);
+	strcpy(wavpath, directory);
 	char* wavname=&wavpath[strlen(wavpath)];
 	
-	for(int i=0;i<num_voice_types;i++)	
+	for (int i=0;i<num_voice_types;i++)	
 	{
 		voice_types[i].category = (ScreamSoundCategory*) PoolAllocateMem( num_voice_cats * sizeof(ScreamSoundCategory));
-		for(int j=0;j<num_voice_cats;j++)
+		for (int j=0;j<num_voice_cats;j++)
 		{
 			ScreamSoundCategory* cat=&voice_types[i].category[j];
 			cat->last_sound = SID_NOSOUND;
@@ -116,7 +116,7 @@ void CharacterSoundEffects::LoadSounds(const char* filename, const char* directo
 				cat->sounds = 0;
 			}
 
-			for(int k=0;k<cat->num_sounds;)
+			for (int k=0;k<cat->num_sounds;)
 			{
 				ScreamSound * sound=&cat->sounds[k];
 
@@ -128,8 +128,8 @@ void CharacterSoundEffects::LoadSounds(const char* filename, const char* directo
 				sound->volume=*(int*)bufpos;
 				bufpos+=4;
 
-				sound->sound_loaded=GetSound(wavpath);
-				if(sound->sound_loaded)
+				sound->sound_loaded = GetSound(wavpath);
+				if (sound->sound_loaded)
 				{
 					k++;
 				}
